Replaced dictionary curly braces with list brackets
I'm not the most pro with python, but when I tried to run this code it kept failing and complaining about the curly brace lists. It looks like curly braces are used to create "dictionary" structures, but brackets are for python lists. I changed them and the code is working again.